Make distinction between "enter" and "edit" modes
Excel and Google Sheets have a notion of "Enter" and "Edit" modes for manipulating a cell's data:

- "Enter" mode is activated when a user starts typing in the active cell. This typing replaces the cell's previous value, and allow the user to quickly navigate to adjacent cells using the arrow keys. This allows you to enter a series of values very quickly.
- "Edit" mode is when you hit the "Enter" key (confusing, no?) to edit the cell's value. It places the text cursor at the end of the cell's content, and arrow keys do not move away from the active cell, but instead move the text cursor around. This mode is intended for more deliberate editing of existing values.

canvas-datagrid previously didn't make a clear distinction, so there was a bug when you started typing in a cell and then attempted to use the arrow keys to move the cursor, which left the cell instead. This commit makes the distinction more explicit.