Fix webcomponent data and schema attributes
`component.attributeChangedCallback` is always called after the Grid initialisation. Thus initialisation of the `datagrid` webcomponent with a schema fails, as args is only respected during the initialisation phase. The bug can be tested like that ```      <canvas-datagrid class="myGridStyle" selectionmode="row" autogenerateschema="false"
        schema='[{"name":"title", "title":"The Title"}, {"name":"description", "title":"The Description"}, {"name":"foo", "title":"The Foo"}]'>
        [
            {"title": "row 1 column 1", "description":"row 1 column 2","foo":"row 1 column 3"},
            {"title": "row 1 column 1", "description":"row 1 column 2","foo":"row 1 column 3"}
        ]
      </canvas-datagrid>```
If schema is not applied the titles are auto generated and all lowercase.