Split original source data from viewed data (#334)
* Add chai assertion library to tests

* Rename applyFilter -> applyFilters

* Add property boundData as alias for originalData

* Add property viewData as alias for data

viewData reflects the data that's painted on screen, after filtering and sorting and any other operations have been applied to the original data.

* Read from viewData instead of data for tests

* Use viewData when touch-selecting area

* Rename applyDataTransforms -> refresh

* Move `refreshFromOrigialData` into `refresh` method

Method is only called once, and does not seem to much beyond copy an array.

* Remove unused data declaration

* Use viewData to draw, in context menu

Tests are passing, but that's only because self.data is stil being assigned in .refresh(). Once we remove that, a bunch of tests in events.js will fail. Tackling that in another commit.

* Mark self.data locations (to investigate)

* Improve findRowScrollTop RangeError

* Use `viewData` in `self.resize()`

* Alter originalData when calling insertColumn

* Refactor getSelectedData to use viewData

Renamed variables for clarity

* Base createRowOrders on originalData instead of viewData

... because viewData maybe filtered and thus shorter

* Only create tempSchema is originalData is empty

* Delete column from originalData

* Compare against viewData in endEdit, not data

* self.scroll refers to viewData instead of self.data

* Use viewData where appropriate

* Ensure all tests are run

For some reason, removing the self.data assignment fails tests in such a way that they are not reported as failed, but simply causing them to be skipped silently.

* Add devServer

* Rename variables for clarity

* Stop setting self.data in refresh(), update comments

* Use originalData in self.rowOrder()

* Rename data -> viewData in draw()

* Rename d -> rowIndex

* Rename r -> rowOrderIndex

* Rename d -> rowData

* Rename rd -> rowData

* Clean up drawRowHeader

* Rename n -> rowOrderIndex

* Rename variables

s -> schema
o -> columnOrderIndex
i -> headerIndex

* Rename n -> rowOrderIndex

* Rename i -> headerIndex

* Rename o -> columnOrderIndex

* Add viewRowIndex/viewColumnIndex to cell

* Add failing tests for viewRowIndex/viewColumnIndex

* Revise failing tests, add `getBoundRowIndexFromViewRowIndex` method

* Rename getBoundRowIndex -> getBoundRowIndexFromViewRowIndex

* Turn the other tests back on

* Calculate viewData based on filters and sorting

Because filtering and sorting changes the view row index, we need to calculate a mapping from viewRowIndex to boundRowIndex (i.e. position in original data array) each time we apply filters or sort.

This commit renames applyFilters, and the new function includes both filtering and sorting to make sure the view index to original index mapping remains correct. We could have split this up into two distinct steps, filtering and sorting, but it would have meant some code duplication for generating the row index mapping, and I'm doubtful it would serve much purpose, as self.orderings.sort() was only called in refresh anyway.

* Call refresh() when re-ordering column

* Fix beforeEndEdit event

* Fix pasteData to alter originalData

* Serve tutorials in development mode

* Add hasActiveFilters method

* Render actual row numbers when filtering

Instead of simply rendering the viewed row index, this commit renders the row numbers as they appear in the original data array.