Fill data on selection handle move
This adds a handle to the desktop version that
users can drag and drop to quickly fill a region.

When the handle is dragged, an overlay will
appear covering the previous selection area and
later expanding upon it when moved. When the user
releases the handle, this fills the region that
the overlay now covers with the data from the
previous selection area.

Similarly to paste events, this can benefit from
'fillCellCallback' to transform the data before
placing it into the cell.